refactor: extract modular mixins for testing manager
