--- src/core/managers/testing_framework_manager.py
@@ -1,753 +1,0 @@
-#!/usr/bin/env python3
-"""
-Testing Framework Manager - V2 Core Manager Consolidation System
-===============================================================
-
-CONSOLIDATED testing framework - replaces multiple separate testing classes with single, specialized manager.
-Consolidates: TDDTestRunner, TestRunner, TestSuiteRunner, TestOrchestrator, TestDataFactory, TestAssertions
-
-Follows V2 standards: OOP design, SRP, no strict LOC limits.
-
-Author: V2 SWARM CAPTAIN
-License: MIT
-"""
-
-import logging
-import json
-import time
-import unittest
-import subprocess
-import sys
-from pathlib import Path
-from typing import Dict, List, Optional, Any, Callable, Union, Type
-from dataclasses import dataclass, asdict
-from datetime import datetime, timedelta
-from unittest import TestCase, TestSuite, TestResult
-
-from ..base_manager import BaseManager, ManagerStatus, ManagerPriority
-
-logger = logging.getLogger(__name__)
-
-
-# CONSOLIDATED TESTING TYPES
-@dataclass
-class TestExecutionResult:
-    """Represents the result of a test execution."""
-    test_name: str
-    test_class: str
-    status: str  # "passed", "failed", "error", "skipped"
-    execution_time: float
-    error_message: Optional[str] = None
-    stack_trace: Optional[str] = None
-    metadata: Dict[str, Any] = None
-
-
-@dataclass
-class TestSuiteResult:
-    """Represents the result of a test suite execution."""
-    suite_name: str
-    total_tests: int
-    passed_tests: int
-    failed_tests: int
-    error_tests: int
-    skipped_tests: int
-    execution_time: float
-    start_time: float
-    end_time: float
-    test_results: List[TestExecutionResult]
-    metadata: Dict[str, Any] = None
-
-
-@dataclass
-class TestConfiguration:
-    """Represents testing framework configuration."""
-    framework_type: str  # "unittest", "pytest", "custom"
-    parallel_execution: bool = False
-    max_workers: int = 4
-    timeout_seconds: int = 300
-    verbose_output: bool = True
-    generate_reports: bool = True
-    coverage_enabled: bool = False
-    retry_failed_tests: bool = False
-    max_retries: int = 3
-
-
-class TestingFrameworkManager(BaseManager):
-    """
-    UNIFIED Testing Framework Manager - Single responsibility: All testing operations
-    
-    This manager consolidates functionality from:
-    - Multiple test runner classes
-    - Test orchestration systems
-    - Test utility classes
-    - Test data factories
-    
-    Total consolidation: Multiple files → 1 file (100% duplication eliminated)
-    """
-
-    def __init__(self, config_path: str = "config/testing_framework_manager.json"):
-        """Initialize unified testing framework manager"""
-        super().__init__(
-            manager_name="TestingFrameworkManager",
-            config_path=config_path,
-            enable_metrics=True,
-            enable_events=True,
-            enable_persistence=True
-        )
-        
-        # Testing framework state
-        self._test_suites: Dict[str, TestSuite] = {}
-        self._test_results: List[TestSuiteResult] = []
-        self._test_configurations: Dict[str, TestConfiguration] = {}
-        self._test_data_factories: Dict[str, Callable] = {}
-        self._test_assertions: Dict[str, Callable] = {}
-        
-        # Performance tracking
-        self._execution_history: List[Dict[str, Any]] = []
-        self._performance_metrics: Dict[str, List[float]] = defaultdict(list)
-        self._coverage_data: Dict[str, Dict[str, float]] = {}
-        
-        # Configuration
-        self.default_framework = "unittest"
-        self.max_parallel_tests = 4
-        self.test_timeout = 300
-        self.retry_failed_tests = True
-        self.max_retries = 3
-        
-        # Initialize testing framework
-        self._load_manager_config()
-        self._initialize_testing_workspace()
-        self._register_default_test_utilities()
-    
-    # SPECIALIZED TESTING FRAMEWORK CAPABILITIES - ENHANCED FOR V2
-    def analyze_testing_performance_patterns(self, time_range_hours: int = 24) -> Dict[str, Any]:
-        """Analyze testing performance patterns for optimization insights"""
-        try:
-            # Get recent performance data
-            recent_time = time.time() - (time_range_hours * 3600)
-            
-            performance_analysis = {
-                "total_test_suites": len(self._test_suites),
-                "total_test_executions": len(self._test_results),
-                "recent_executions": 0,
-                "performance_trends": {},
-                "bottleneck_analysis": {},
-                "optimization_opportunities": [],
-                "coverage_metrics": {}
-            }
-            
-            # Analyze recent test executions
-            recent_results = [r for r in self._test_results if r.start_time > recent_time]
-            performance_analysis["recent_executions"] = len(recent_results)
-            
-            if recent_results:
-                # Performance trends
-                execution_times = [r.execution_time for r in recent_results]
-                performance_analysis["performance_trends"] = {
-                    "average_execution_time": sum(execution_times) / len(execution_times),
-                    "min_execution_time": min(execution_times),
-                    "max_execution_time": max(execution_times),
-                    "execution_time_variance": self._calculate_variance(execution_times)
-                }
-                
-                # Success rates
-                total_tests = sum(r.total_tests for r in recent_results)
-                total_passed = sum(r.passed_tests for r in recent_results)
-                total_failed = sum(r.failed_tests for r in recent_results)
-                
-                performance_analysis["success_metrics"] = {
-                    "total_tests": total_tests,
-                    "passed_tests": total_passed,
-                    "failed_tests": total_failed,
-                    "success_rate": total_passed / total_tests if total_tests > 0 else 0,
-                    "failure_rate": total_failed / total_tests if total_tests > 0 else 0
-                }
-                
-                # Identify bottlenecks
-                if performance_analysis["performance_trends"]["average_execution_time"] > 60:
-                    performance_analysis["bottleneck_analysis"]["slow_execution"] = {
-                        "issue": "Slow test execution detected",
-                        "severity": "high",
-                        "recommendation": "Optimize test setup/teardown or reduce test complexity"
-                    }
-                
-                if performance_analysis["success_metrics"]["failure_rate"] > 0.1:
-                    performance_analysis["bottleneck_analysis"]["high_failure_rate"] = {
-                        "issue": "High test failure rate detected",
-                        "severity": "critical",
-                        "recommendation": "Investigate test stability and environment issues"
-                    }
-                
-                # Coverage analysis
-                if self._coverage_data:
-                    performance_analysis["coverage_metrics"] = {
-                        "average_coverage": sum(sum(c.values()) / len(c) for c in self._coverage_data.values()) / len(self._coverage_data),
-                        "coverage_distribution": {k: sum(v.values()) / len(v) for k, v in self._coverage_data.items()}
-                    }
-            
-            # Generate optimization opportunities
-            if performance_analysis.get("bottleneck_analysis"):
-                for bottleneck_type, details in performance_analysis["bottleneck_analysis"].items():
-                    performance_analysis["optimization_opportunities"].append(details["recommendation"])
-            
-            logger.info(f"Testing performance analysis completed")
-            return performance_analysis
-            
-        except Exception as e:
-            logger.error(f"Failed to analyze testing performance patterns: {e}")
-            return {"error": str(e)}
-    
-    def create_intelligent_testing_strategy(self, strategy_type: str, parameters: Dict[str, Any]) -> str:
-        """Create an intelligent testing strategy with adaptive parameters"""
-        try:
-            strategy_id = f"intelligent_testing_{strategy_type}_{int(time.time())}"
-            
-            if strategy_type == "adaptive_test_execution":
-                strategy_config = {
-                    "id": strategy_id,
-                    "type": "adaptive_test_execution",
-                    "description": "Dynamically adjust test execution based on performance patterns and resource availability",
-                    "parameters": {
-                        **parameters,
-                        "performance_threshold": parameters.get("performance_threshold", 0.8),
-                        "resource_optimization": parameters.get("resource_optimization", True),
-                        "adaptive_timeout": parameters.get("adaptive_timeout", True)
-                    }
-                }
-                
-            elif strategy_type == "intelligent_test_prioritization":
-                strategy_config = {
-                    "id": strategy_id,
-                    "type": "intelligent_test_prioritization",
-                    "description": "Prioritize tests based on failure probability, execution time, and business impact",
-                    "parameters": {
-                        **parameters,
-                        "failure_probability_weight": parameters.get("failure_probability_weight", 0.4),
-                        "execution_time_weight": parameters.get("execution_time_weight", 0.3),
-                        "business_impact_weight": parameters.get("business_impact_weight", 0.3)
-                    }
-                }
-                
-            elif strategy_type == "coverage_optimization":
-                strategy_config = {
-                    "id": strategy_id,
-                    "type": "coverage_optimization",
-                    "description": "Optimize test coverage by identifying gaps and prioritizing high-impact test cases",
-                    "parameters": {
-                        **parameters,
-                        "coverage_threshold": parameters.get("coverage_threshold", 0.8),
-                        "gap_analysis": parameters.get("gap_analysis", True),
-                        "impact_prioritization": parameters.get("impact_prioritization", True)
-                    }
-                }
-                
-            else:
-                raise ValueError(f"Unknown testing strategy type: {strategy_type}")
-            
-            # Store strategy configuration
-            if not hasattr(self, 'intelligent_strategies'):
-                self.intelligent_strategies = {}
-            self.intelligent_strategies[strategy_id] = strategy_config
-            
-            logger.info(f"Created intelligent testing strategy: {strategy_id}")
-            return strategy_id
-            
-        except Exception as e:
-            logger.error(f"Failed to create intelligent testing strategy: {e}")
-            raise
-    
-    def execute_intelligent_testing_strategy(self, strategy_id: str, context: Dict[str, Any]) -> Dict[str, Any]:
-        """Execute intelligent testing strategy"""
-        try:
-            if not hasattr(self, 'intelligent_strategies') or strategy_id not in self.intelligent_strategies:
-                raise ValueError(f"Strategy configuration not found: {strategy_id}")
-            
-            strategy_config = self.intelligent_strategies[strategy_id]
-            strategy_type = strategy_config["type"]
-            
-            execution_result = {
-                "strategy_id": strategy_id,
-                "strategy_type": strategy_type,
-                "actions_taken": [],
-                "performance_impact": {},
-                "recommendations": []
-            }
-            
-            if strategy_type == "adaptive_test_execution":
-                # Execute adaptive test execution
-                execution_result.update(self._execute_adaptive_test_execution(strategy_config, context))
-                
-            elif strategy_type == "intelligent_test_prioritization":
-                # Execute intelligent test prioritization
-                execution_result.update(self._execute_intelligent_test_prioritization(strategy_config, context))
-                
-            elif strategy_type == "coverage_optimization":
-                # Execute coverage optimization
-                execution_result.update(self._execute_coverage_optimization(strategy_config, context))
-            
-            logger.info(f"Intelligent testing strategy executed: {strategy_id}")
-            return execution_result
-            
-        except Exception as e:
-            logger.error(f"Failed to execute intelligent testing strategy: {e}")
-            raise
-    
-    def predict_testing_needs(self, time_horizon_minutes: int = 30) -> List[Dict[str, Any]]:
-        """Predict potential testing needs based on current patterns"""
-        try:
-            predictions = []
-            performance_analysis = self.analyze_testing_performance_patterns(time_horizon_minutes / 60)
-            
-            # Check for performance degradation
-            performance_trends = performance_analysis.get("performance_trends", {})
-            if performance_trends.get("average_execution_time", 0) > 120:
-                prediction = {
-                    "issue_type": "performance_degradation",
-                    "probability": 0.8,
-                    "estimated_time_to_threshold": time_horizon_minutes * 0.3,
-                    "severity": "high",
-                    "recommended_action": "Optimize test execution or reduce test load"
-                }
-                predictions.append(prediction)
-            
-            # Check for coverage gaps
-            coverage_metrics = performance_analysis.get("coverage_metrics", {})
-            if coverage_metrics.get("average_coverage", 1.0) < 0.7:
-                prediction = {
-                    "issue_type": "coverage_gap",
-                    "probability": 0.9,
-                    "estimated_time_to_threshold": time_horizon_minutes * 0.5,
-                    "severity": "medium",
-                    "recommended_action": "Add test cases to improve coverage"
-                }
-                predictions.append(prediction)
-            
-            # Check for resource pressure
-            if len(self._test_suites) > 50:
-                prediction = {
-                    "issue_type": "resource_pressure",
-                    "probability": 0.7,
-                    "estimated_time_to_threshold": time_horizon_minutes * 0.8,
-                    "severity": "medium",
-                    "recommended_action": "Consider test suite consolidation or parallel execution"
-                }
-                predictions.append(prediction)
-            
-            logger.info(f"Testing needs prediction completed: {len(predictions)} predictions identified")
-            return predictions
-            
-        except Exception as e:
-            logger.error(f"Failed to predict testing needs: {e}")
-            return []
-    
-    def optimize_testing_operations_automatically(self) -> Dict[str, Any]:
-        """Automatically optimize testing operations based on current patterns"""
-        try:
-            optimization_plan = {
-                "optimizations_applied": [],
-                "performance_improvements": {},
-                "recommendations": []
-            }
-            
-            # Analyze current testing state
-            performance_analysis = self.analyze_testing_performance_patterns()
-            
-            # Apply automatic optimizations
-            performance_trends = performance_analysis.get("performance_trends", {})
-            if performance_trends.get("average_execution_time", 0) > 60:
-                # Slow execution - optimize test configuration
-                self._optimize_test_configuration()
-                optimization_plan["optimizations_applied"].append({
-                    "action": "test_configuration_optimization",
-                    "target": "execution_time < 60s",
-                    "status": "executed"
-                })
-                optimization_plan["performance_improvements"]["execution_time"] = "optimized"
-            
-            # Check for coverage optimization opportunities
-            coverage_metrics = performance_analysis.get("coverage_metrics", {})
-            if coverage_metrics.get("average_coverage", 1.0) < 0.8:
-                # Low coverage - identify gaps
-                gaps = self._identify_coverage_gaps()
-                optimization_plan["optimizations_applied"].append({
-                    "action": "coverage_gap_analysis",
-                    "target": "coverage > 80%",
-                    "status": "executed"
-                })
-                optimization_plan["performance_improvements"]["coverage_analysis"] = "completed"
-            
-            # Generate recommendations
-            if not optimization_plan["optimizations_applied"]:
-                optimization_plan["recommendations"].append("Testing operations are optimized")
-            else:
-                optimization_plan["recommendations"].append("Monitor optimization results for 15 minutes")
-                optimization_plan["recommendations"].append("Consider implementing permanent optimizations")
-            
-            logger.info(f"Automatic testing optimization completed: {len(optimization_plan['optimizations_applied'])} optimizations applied")
-            return optimization_plan
-            
-        except Exception as e:
-            logger.error(f"Failed to optimize testing operations automatically: {e}")
-            return {"error": str(e)}
-    
-    def generate_testing_report(self, report_type: str = "comprehensive") -> Dict[str, Any]:
-        """Generate comprehensive testing framework report"""
-        try:
-            report = {
-                "report_id": f"testing_framework_report_{int(time.time())}",
-                "generated_at": datetime.now().isoformat(),
-                "report_type": report_type,
-                "summary": {},
-                "detailed_metrics": {},
-                "test_suite_summary": {},
-                "recommendations": []
-            }
-            
-            # Generate summary
-            total_suites = len(self._test_suites)
-            total_executions = len(self._test_results)
-            total_tests = sum(r.total_tests for r in self._test_results) if self._test_results else 0
-            total_passed = sum(r.passed_tests for r in self._test_results) if self._test_results else 0
-            
-            report["summary"] = {
-                "total_test_suites": total_suites,
-                "total_test_executions": total_executions,
-                "total_tests_executed": total_tests,
-                "total_tests_passed": total_passed,
-                "overall_success_rate": total_passed / total_tests if total_tests > 0 else 0,
-                "framework_status": self.status.value
-            }
-            
-            # Generate detailed metrics
-            if self._test_results:
-                recent_results = self._test_results[-10:]  # Last 10 executions
-                report["detailed_metrics"] = {
-                    "recent_executions": len(recent_results),
-                    "average_execution_time": sum(r.execution_time for r in recent_results) / len(recent_results),
-                    "success_rate_trend": [r.passed_tests / r.total_tests for r in recent_results if r.total_tests > 0],
-                    "coverage_trend": list(self._coverage_data.values())[-10:] if self._coverage_data else []
-                }
-            
-            # Generate test suite summary
-            if self._test_suites:
-                suite_sizes = {name: len(suite) for name, suite in self._test_suites.items()}
-                report["test_suite_summary"] = {
-                    "largest_test_suites": sorted(suite_sizes.items(), key=lambda x: x[1], reverse=True)[:5],
-                    "smallest_test_suites": sorted(suite_sizes.items(), key=lambda x: x[1])[:5],
-                    "suite_size_distribution": suite_sizes
-                }
-            
-            # Generate recommendations
-            performance_analysis = self.analyze_testing_performance_patterns()
-            for opportunity in performance_analysis.get("optimization_opportunities", []):
-                report["recommendations"].append(opportunity)
-            
-            # Check for testing efficiency
-            if total_tests > 0 and total_executions > 0:
-                avg_tests_per_execution = total_tests / total_executions
-                if avg_tests_per_execution > 100:
-                    report["recommendations"].append("High tests per execution - consider splitting large test suites")
-                elif avg_tests_per_execution < 5:
-                    report["recommendations"].append("Low tests per execution - consider consolidating small test suites")
-            
-            logger.info(f"Testing framework report generated: {report['report_id']}")
-            return report
-            
-        except Exception as e:
-            logger.error(f"Failed to generate testing framework report: {e}")
-            return {"error": str(e)}
-    
-    # TEST EXECUTION METHODS (from various TestRunner classes)
-    def run_test_suite(self, suite_name: str, test_classes: List[Type[TestCase]] = None) -> TestSuiteResult:
-        """Run a complete test suite."""
-        try:
-            start_time = time.time()
-            
-            # Create or get test suite
-            if suite_name not in self._test_suites:
-                if test_classes:
-                    suite = unittest.TestSuite()
-                    for test_class in test_classes:
-                        suite.addTests(unittest.TestLoader().loadTestsFromTestCase(test_class))
-                    self._test_suites[suite_name] = suite
-                else:
-                    raise ValueError(f"Test suite '{suite_name}' not found and no test classes provided")
-            
-            suite = self._test_suites[suite_name]
-            
-            # Execute tests
-            runner = unittest.TextTestRunner(verbosity=2)
-            result = runner.run(suite)
-            
-            end_time = time.time()
-            execution_time = end_time - start_time
-            
-            # Create test suite result
-            suite_result = TestSuiteResult(
-                suite_name=suite_name,
-                total_tests=result.testsRun,
-                passed_tests=result.testsRun - len(result.failures) - len(result.errors),
-                failed_tests=len(result.failures),
-                error_tests=len(result.errors),
-                skipped_tests=len(result.skipped) if hasattr(result, 'skipped') else 0,
-                execution_time=execution_time,
-                start_time=start_time,
-                end_time=end_time,
-                test_results=[],
-                metadata={"runner_type": "unittest.TextTestRunner"}
-            )
-            
-            # Process individual test results
-            for test, traceback in result.failures:
-                test_result = TestExecutionResult(
-                    test_name=test._testMethodName,
-                    test_class=test.__class__.__name__,
-                    status="failed",
-                    execution_time=0.0,  # Individual timing not available
-                    error_message=str(traceback),
-                    metadata={"test_id": str(test)}
-                )
-                suite_result.test_results.append(test_result)
-            
-            for test, traceback in result.errors:
-                test_result = TestExecutionResult(
-                    test_name=test._testMethodName,
-                    test_class=test.__class__.__name__,
-                    status="error",
-                    execution_time=0.0,
-                    error_message=str(traceback),
-                    metadata={"test_id": str(test)}
-                )
-                suite_result.test_results.append(test_result)
-            
-            # Store result
-            self._test_results.append(suite_result)
-            
-            # Record performance metrics
-            self._performance_metrics[suite_name].append(execution_time)
-            
-            logger.info(f"Test suite '{suite_name}' executed: {suite_result.passed_tests}/{suite_result.total_tests} passed")
-            return suite_result
-            
-        except Exception as e:
-            logger.error(f"Failed to run test suite '{suite_name}': {e}")
-            raise
-    
-    def run_tdd_tests(self, test_categories: List[str] = None) -> Dict[str, TestSuiteResult]:
-        """Run TDD tests with category-based organization."""
-        try:
-            if not test_categories:
-                test_categories = ["unit", "integration", "smoke"]
-            
-            results = {}
-            
-            for category in test_categories:
-                # Find tests in category
-                category_tests = self._find_tests_by_category(category)
-                if category_tests:
-                    suite_name = f"tdd_{category}_tests"
-                    result = self.run_test_suite(suite_name, category_tests)
-                    results[category] = result
-            
-            logger.info(f"TDD tests executed for categories: {list(results.keys())}")
-            return results
-            
-        except Exception as e:
-            logger.error(f"Failed to run TDD tests: {e}")
-            raise
-    
-    def run_smoke_tests(self) -> TestSuiteResult:
-        """Run smoke tests for basic system validation."""
-        try:
-            smoke_tests = self._find_smoke_tests()
-            if smoke_tests:
-                return self.run_test_suite("smoke_tests", smoke_tests)
-            else:
-                logger.warning("No smoke tests found")
-                return None
-                
-        except Exception as e:
-            logger.error(f"Failed to run smoke tests: {e}")
-            raise
-    
-    # TEST UTILITY METHODS (from TestDataFactory and TestAssertions)
-    def create_test_data(self, data_type: str, **kwargs) -> Any:
-        """Create test data using registered factories."""
-        try:
-            if data_type in self._test_data_factories:
-                factory = self._test_data_factories[data_type]
-                return factory(**kwargs)
-            else:
-                logger.warning(f"No test data factory found for type: {data_type}")
-                return None
-                
-        except Exception as e:
-            logger.error(f"Failed to create test data for type '{data_type}': {e}")
-            return None
-    
-    def register_test_data_factory(self, data_type: str, factory_func: Callable) -> None:
-        """Register a test data factory function."""
-        self._test_data_factories[data_type] = factory_func
-        logger.info(f"Registered test data factory for type: {data_type}")
-    
-    def assert_test_condition(self, assertion_type: str, *args, **kwargs) -> bool:
-        """Execute test assertion using registered assertion functions."""
-        try:
-            if assertion_type in self._test_assertions:
-                assertion_func = self._test_assertions[assertion_type]
-                return assertion_func(*args, **kwargs)
-            else:
-                logger.warning(f"No test assertion found for type: {assertion_type}")
-                return False
-                
-        except Exception as e:
-            logger.error(f"Failed to execute test assertion '{assertion_type}': {e}")
-            return False
-    
-    def register_test_assertion(self, assertion_type: str, assertion_func: Callable) -> None:
-        """Register a test assertion function."""
-        self._test_assertions[assertion_type] = assertion_func
-        logger.info(f"Registered test assertion for type: {assertion_type}")
-    
-    # TEST ORCHESTRATION METHODS (from TestOrchestrator)
-    def orchestrate_test_execution(self, execution_plan: Dict[str, Any]) -> Dict[str, TestSuiteResult]:
-        """Orchestrate complex test execution based on plan."""
-        try:
-            results = {}
-            
-            # Parse execution plan
-            test_suites = execution_plan.get("test_suites", [])
-            execution_order = execution_plan.get("execution_order", "sequential")
-            parallel_suites = execution_plan.get("parallel_suites", [])
-            
-            if execution_order == "parallel" and parallel_suites:
-                # Execute parallel suites first
-                for suite_name in parallel_suites:
-                    if suite_name in self._test_suites:
-                        # In a real implementation, this would use threading/multiprocessing
-                        result = self.run_test_suite(suite_name)
-                        results[suite_name] = result
-            
-            # Execute remaining suites sequentially
-            for suite_name in test_suites:
-                if suite_name not in results and suite_name in self._test_suites:
-                    result = self.run_test_suite(suite_name)
-                    results[suite_name] = result
-            
-            logger.info(f"Test orchestration completed: {len(results)} suites executed")
-            return results
-            
-        except Exception as e:
-            logger.error(f"Failed to orchestrate test execution: {e}")
-            raise
-    
-    # STRATEGY EXECUTION METHODS
-    def _execute_adaptive_test_execution(self, strategy_config: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Execute adaptive test execution strategy"""
-        # Simplified implementation
-        return {
-            "actions_taken": ["test_execution_optimization"],
-            "performance_impact": {"execution_time": "optimized"},
-            "recommendations": ["Monitor test execution performance for 15 minutes"]
-        }
-    
-    def _execute_intelligent_test_prioritization(self, strategy_config: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Execute intelligent test prioritization strategy"""
-        # Simplified implementation
-        return {
-            "actions_taken": ["test_prioritization_optimization"],
-            "performance_impact": {"test_ordering": "optimized"},
-            "recommendations": ["Review test priority assignments"]
-        }
-    
-    def _execute_coverage_optimization(self, strategy_config: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Execute coverage optimization strategy"""
-        # Simplified implementation
-        return {
-            "actions_taken": ["coverage_optimization"],
-            "performance_impact": {"coverage": "improved"},
-            "recommendations": ["Monitor coverage metrics"]
-        }
-    
-    # UTILITY METHODS
-    def _load_manager_config(self):
-        """Load manager-specific configuration"""
-        try:
-            if Path(self.config_path).exists():
-                with open(self.config_path, 'r') as f:
-                    config = json.load(f)
-                    # Load testing-specific configuration
-                    if "testing" in config:
-                        testing_config = config["testing"]
-                        self.default_framework = testing_config.get("default_framework", "unittest")
-                        self.max_parallel_tests = testing_config.get("max_parallel_tests", 4)
-                        self.test_timeout = testing_config.get("test_timeout", 300)
-                        self.retry_failed_tests = testing_config.get("retry_failed_tests", True)
-                        self.max_retries = testing_config.get("max_retries", 3)
-            else:
-                logger.warning(f"Testing config file not found: {self.config_path}")
-        except Exception as e:
-            logger.error(f"Failed to load testing config: {e}")
-    
-    def _initialize_testing_workspace(self):
-        """Initialize testing workspace"""
-        self.workspace_path = Path("testing_workspaces")
-        self.workspace_path.mkdir(exist_ok=True)
-        logger.info("Testing workspace initialized")
-    
-    def _register_default_test_utilities(self):
-        """Register default test utilities"""
-        # Register default test data factories
-        self.register_test_data_factory("string", lambda length=10: "a" * length)
-        self.register_test_data_factory("integer", lambda min_val=0, max_val=100: random.randint(min_val, max_val))
-        self.register_test_data_factory("list", lambda size=5: list(range(size)))
-        
-        # Register default test assertions
-        self.register_test_assertion("equals", lambda a, b: a == b)
-        self.register_test_assertion("not_equals", lambda a, b: a != b)
-        self.register_test_assertion("contains", lambda container, item: item in container)
-        self.register_test_assertion("greater_than", lambda a, b: a > b)
-        self.register_test_assertion("less_than", lambda a, b: a < b)
-        
-        logger.info("Default test utilities registered")
-    
-    def _find_tests_by_category(self, category: str) -> List[Type[TestCase]]:
-        """Find test classes by category."""
-        # Simplified implementation - in reality, this would scan test files
-        return []
-    
-    def _find_smoke_tests(self) -> List[Type[TestCase]]:
-        """Find smoke test classes."""
-        # Simplified implementation - in reality, this would scan test files
-        return []
-    
-    def _optimize_test_configuration(self) -> None:
-        """Optimize test configuration for better performance."""
-        # Simplified optimization logic
-        logger.info("Test configuration optimized")
-    
-    def _identify_coverage_gaps(self) -> List[Dict[str, Any]]:
-        """Identify coverage gaps in test suite."""
-        # Simplified gap analysis
-        return []
-    
-    def _calculate_variance(self, values: List[float]) -> float:
-        """Calculate variance of a list of values."""
-        if not values:
-            return 0.0
-        mean = sum(values) / len(values)
-        return sum((x - mean) ** 2 for x in values) / len(values)
-    
-    def cleanup(self):
-        """Cleanup testing framework manager resources"""
-        try:
-            # Clear test suites and results
-            self._test_suites.clear()
-            self._test_results.clear()
-            logger.info("TestingFrameworkManager cleanup completed")
-        except Exception as e:
-            logger.error(f"TestingFrameworkManager cleanup failed: {e}")
-
-
-# Import random for test data generation
-import random
