+++ src/core/managers/testing_framework_manager/performance_analysis.py
@@ -1,0 +1,310 @@
+#!/usr/bin/env python3
+"""Performance analysis utilities for the testing framework manager."""
+
+import logging
+import time
+from datetime import datetime
+from typing import Any, Dict, List
+
+logger = logging.getLogger(__name__)
+
+
+class PerformanceAnalysisMixin:
+    """Mixin providing performance analysis and reporting capabilities."""
+
+    def analyze_testing_performance_patterns(
+        self, time_range_hours: int = 24
+    ) -> Dict[str, Any]:
+        """Analyze testing performance patterns for optimization insights."""
+        try:
+            recent_time = time.time() - (time_range_hours * 3600)
+
+            performance_analysis = {
+                "total_test_suites": len(self._test_suites),
+                "total_test_executions": len(self._test_results),
+                "recent_executions": 0,
+                "performance_trends": {},
+                "bottleneck_analysis": {},
+                "optimization_opportunities": [],
+                "coverage_metrics": {},
+            }
+
+            recent_results = [
+                r for r in self._test_results if r.start_time > recent_time
+            ]
+            performance_analysis["recent_executions"] = len(recent_results)
+
+            if recent_results:
+                execution_times = [r.execution_time for r in recent_results]
+                performance_analysis["performance_trends"] = {
+                    "average_execution_time": sum(execution_times)
+                    / len(execution_times),
+                    "min_execution_time": min(execution_times),
+                    "max_execution_time": max(execution_times),
+                    "execution_time_variance": self._calculate_variance(
+                        execution_times
+                    ),
+                }
+
+                total_tests = sum(r.total_tests for r in recent_results)
+                total_passed = sum(r.passed_tests for r in recent_results)
+                total_failed = sum(r.failed_tests for r in recent_results)
+                performance_analysis["success_metrics"] = {
+                    "total_tests": total_tests,
+                    "passed_tests": total_passed,
+                    "failed_tests": total_failed,
+                    "success_rate": total_passed / total_tests
+                    if total_tests > 0
+                    else 0,
+                    "failure_rate": total_failed / total_tests
+                    if total_tests > 0
+                    else 0,
+                }
+
+                if (
+                    performance_analysis["performance_trends"]["average_execution_time"]
+                    > 60
+                ):
+                    performance_analysis["bottleneck_analysis"]["slow_execution"] = {
+                        "issue": "Slow test execution detected",
+                        "severity": "high",
+                        "recommendation": "Optimize test setup/teardown or reduce test complexity",
+                    }
+
+                if performance_analysis["success_metrics"]["failure_rate"] > 0.1:
+                    performance_analysis["bottleneck_analysis"]["high_failure_rate"] = {
+                        "issue": "High test failure rate detected",
+                        "severity": "critical",
+                        "recommendation": "Investigate test stability and environment issues",
+                    }
+
+                if self._coverage_data:
+                    performance_analysis["coverage_metrics"] = {
+                        "average_coverage": sum(
+                            sum(c.values()) / len(c)
+                            for c in self._coverage_data.values()
+                        )
+                        / len(self._coverage_data),
+                        "coverage_distribution": {
+                            k: sum(v.values()) / len(v)
+                            for k, v in self._coverage_data.items()
+                        },
+                    }
+
+            if performance_analysis.get("bottleneck_analysis"):
+                for details in performance_analysis["bottleneck_analysis"].values():
+                    performance_analysis["optimization_opportunities"].append(
+                        details["recommendation"]
+                    )
+
+            logger.info("Testing performance analysis completed")
+            return performance_analysis
+
+        except Exception as e:
+            logger.error(f"Failed to analyze testing performance patterns: {e}")
+            return {"error": str(e)}
+
+    def predict_testing_needs(
+        self, time_horizon_minutes: int = 30
+    ) -> List[Dict[str, Any]]:
+        """Predict potential testing needs based on current patterns."""
+        try:
+            predictions: List[Dict[str, Any]] = []
+            performance_analysis = self.analyze_testing_performance_patterns(
+                time_horizon_minutes / 60
+            )
+
+            performance_trends = performance_analysis.get("performance_trends", {})
+            if performance_trends.get("average_execution_time", 0) > 120:
+                predictions.append(
+                    {
+                        "issue_type": "performance_degradation",
+                        "probability": 0.8,
+                        "estimated_time_to_threshold": time_horizon_minutes * 0.3,
+                        "severity": "high",
+                        "recommended_action": "Optimize test execution or reduce test load",
+                    }
+                )
+
+            coverage_metrics = performance_analysis.get("coverage_metrics", {})
+            if coverage_metrics.get("average_coverage", 1.0) < 0.7:
+                predictions.append(
+                    {
+                        "issue_type": "coverage_gap",
+                        "probability": 0.9,
+                        "estimated_time_to_threshold": time_horizon_minutes * 0.5,
+                        "severity": "medium",
+                        "recommended_action": "Add test cases to improve coverage",
+                    }
+                )
+
+            if len(self._test_suites) > 50:
+                predictions.append(
+                    {
+                        "issue_type": "resource_pressure",
+                        "probability": 0.7,
+                        "estimated_time_to_threshold": time_horizon_minutes * 0.8,
+                        "severity": "medium",
+                        "recommended_action": "Consider test suite consolidation or parallel execution",
+                    }
+                )
+
+            logger.info(
+                "Testing needs prediction completed: %d predictions identified",
+                len(predictions),
+            )
+            return predictions
+
+        except Exception as e:
+            logger.error(f"Failed to predict testing needs: {e}")
+            return []
+
+    def optimize_testing_operations_automatically(self) -> Dict[str, Any]:
+        """Automatically optimize testing operations based on current patterns."""
+        try:
+            optimization_plan = {
+                "optimizations_applied": [],
+                "performance_improvements": {},
+                "recommendations": [],
+            }
+
+            performance_analysis = self.analyze_testing_performance_patterns()
+            performance_trends = performance_analysis.get("performance_trends", {})
+            if performance_trends.get("average_execution_time", 0) > 60:
+                self._optimize_test_configuration()
+                optimization_plan["optimizations_applied"].append(
+                    {
+                        "action": "test_configuration_optimization",
+                        "target": "execution_time < 60s",
+                        "status": "executed",
+                    }
+                )
+                optimization_plan["performance_improvements"][
+                    "execution_time"
+                ] = "optimized"
+
+            coverage_metrics = performance_analysis.get("coverage_metrics", {})
+            if coverage_metrics.get("average_coverage", 1.0) < 0.8:
+                self._identify_coverage_gaps()
+                optimization_plan["optimizations_applied"].append(
+                    {
+                        "action": "coverage_gap_analysis",
+                        "target": "coverage > 80%",
+                        "status": "executed",
+                    }
+                )
+            return optimization_plan
+
+        except Exception as e:
+            logger.error(f"Failed to optimize testing operations: {e}")
+            return {"error": str(e)}
+
+    def generate_testing_report(
+        self, report_type: str = "comprehensive"
+    ) -> Dict[str, Any]:
+        """Generate comprehensive testing framework report."""
+        try:
+            report: Dict[str, Any] = {
+                "report_id": f"testing_framework_report_{int(time.time())}",
+                "generated_at": datetime.now().isoformat(),
+                "report_type": report_type,
+                "summary": {},
+                "detailed_metrics": {},
+                "test_suite_summary": {},
+                "recommendations": [],
+            }
+
+            total_suites = len(self._test_suites)
+            total_executions = len(self._test_results)
+            total_tests = (
+                sum(r.total_tests for r in self._test_results)
+                if self._test_results
+                else 0
+            )
+            total_passed = (
+                sum(r.passed_tests for r in self._test_results)
+                if self._test_results
+                else 0
+            )
+
+            report["summary"] = {
+                "total_test_suites": total_suites,
+                "total_test_executions": total_executions,
+                "total_tests_executed": total_tests,
+                "total_tests_passed": total_passed,
+                "overall_success_rate": total_passed / total_tests
+                if total_tests > 0
+                else 0,
+                "framework_status": self.status.value,
+            }
+
+            if self._test_results:
+                recent_results = self._test_results[-10:]
+                report["detailed_metrics"] = {
+                    "recent_executions": len(recent_results),
+                    "average_execution_time": sum(
+                        r.execution_time for r in recent_results
+                    )
+                    / len(recent_results),
+                    "success_rate_trend": [
+                        r.passed_tests / r.total_tests
+                        for r in recent_results
+                        if r.total_tests > 0
+                    ],
+                    "coverage_trend": list(self._coverage_data.values())[-10:]
+                    if self._coverage_data
+                    else [],
+                }
+
+            if self._test_suites:
+                suite_sizes = {
+                    name: len(suite) for name, suite in self._test_suites.items()
+                }
+                report["test_suite_summary"] = {
+                    "largest_test_suites": sorted(
+                        suite_sizes.items(), key=lambda x: x[1], reverse=True
+                    )[:5],
+                    "smallest_test_suites": sorted(
+                        suite_sizes.items(), key=lambda x: x[1]
+                    )[:5],
+                    "suite_size_distribution": suite_sizes,
+                }
+
+            performance_analysis = self.analyze_testing_performance_patterns()
+            for opportunity in performance_analysis.get(
+                "optimization_opportunities", []
+            ):
+                report["recommendations"].append(opportunity)
+
+            if total_tests > 0 and total_executions > 0:
+                avg_tests_per_execution = total_tests / total_executions
+                if avg_tests_per_execution > 100:
+                    report["recommendations"].append(
+                        "High tests per execution - consider splitting large test suites"
+                    )
+                elif avg_tests_per_execution < 5:
+                    report["recommendations"].append(
+                        "Low tests per execution - consider consolidating small test suites"
+                    )
+
+            logger.info("Testing framework report generated: %s", report["report_id"])
+            return report
+
+        except Exception as e:
+            logger.error(f"Failed to generate testing framework report: {e}")
+            return {"error": str(e)}
+
+    def _optimize_test_configuration(self) -> None:
+        """Optimize test configuration for better performance."""
+        logger.info("Test configuration optimized")
+
+    def _identify_coverage_gaps(self) -> List[Dict[str, Any]]:
+        """Identify coverage gaps in test suite."""
+        return []
+
+    def _calculate_variance(self, values: List[float]) -> float:
+        """Calculate variance of a list of values."""
+        if not values:
+            return 0.0
+        mean = sum(values) / len(values)
+        return sum((x - mean) ** 2 for x in values) / len(values)
