+++ src/core/managers/testing_framework_manager/strategy_management.py
@@ -1,0 +1,163 @@
+#!/usr/bin/env python3
+"""Intelligent testing strategy management for the framework manager."""
+
+import logging
+import time
+from typing import Any, Dict
+
+logger = logging.getLogger(__name__)
+
+
+class StrategyManagementMixin:
+    """Mixin providing intelligent strategy creation and execution."""
+
+    def create_intelligent_testing_strategy(
+        self, strategy_type: str, parameters: Dict[str, Any]
+    ) -> str:
+        """Create an intelligent testing strategy with adaptive parameters."""
+        try:
+            strategy_id = f"intelligent_testing_{strategy_type}_{int(time.time())}"
+
+            if strategy_type == "adaptive_test_execution":
+                strategy_config = {
+                    "id": strategy_id,
+                    "type": "adaptive_test_execution",
+                    "description": (
+                        "Dynamically adjust test execution based on performance patterns "
+                        "and resource availability"
+                    ),
+                    "parameters": {
+                        **parameters,
+                        "performance_threshold": parameters.get(
+                            "performance_threshold", 0.8
+                        ),
+                        "resource_optimization": parameters.get(
+                            "resource_optimization", True
+                        ),
+                        "adaptive_timeout": parameters.get("adaptive_timeout", True),
+                    },
+                }
+            elif strategy_type == "intelligent_test_prioritization":
+                strategy_config = {
+                    "id": strategy_id,
+                    "type": "intelligent_test_prioritization",
+                    "description": (
+                        "Prioritize tests based on failure probability, execution time, "
+                        "and business impact"
+                    ),
+                    "parameters": {
+                        **parameters,
+                        "failure_probability_weight": parameters.get(
+                            "failure_probability_weight", 0.4
+                        ),
+                        "execution_time_weight": parameters.get(
+                            "execution_time_weight", 0.3
+                        ),
+                        "business_impact_weight": parameters.get(
+                            "business_impact_weight", 0.3
+                        ),
+                    },
+                }
+            elif strategy_type == "coverage_optimization":
+                strategy_config = {
+                    "id": strategy_id,
+                    "type": "coverage_optimization",
+                    "description": (
+                        "Optimize test coverage by identifying gaps and prioritizing "
+                        "high-impact test cases"
+                    ),
+                    "parameters": {
+                        **parameters,
+                        "coverage_threshold": parameters.get("coverage_threshold", 0.8),
+                        "gap_analysis": parameters.get("gap_analysis", True),
+                        "impact_prioritization": parameters.get(
+                            "impact_prioritization", True
+                        ),
+                    },
+                }
+            else:
+                raise ValueError(f"Unknown testing strategy type: {strategy_type}")
+
+            if not hasattr(self, "intelligent_strategies"):
+                self.intelligent_strategies = {}
+            self.intelligent_strategies[strategy_id] = strategy_config
+            logger.info("Created intelligent testing strategy: %s", strategy_id)
+            return strategy_id
+
+        except Exception as e:
+            logger.error(f"Failed to create intelligent testing strategy: {e}")
+            raise
+
+    def execute_intelligent_testing_strategy(
+        self, strategy_id: str, context: Dict[str, Any]
+    ) -> Dict[str, Any]:
+        """Execute an intelligent testing strategy."""
+        try:
+            if (
+                not hasattr(self, "intelligent_strategies")
+                or strategy_id not in self.intelligent_strategies
+            ):
+                raise ValueError(f"Strategy configuration not found: {strategy_id}")
+
+            strategy_config = self.intelligent_strategies[strategy_id]
+            strategy_type = strategy_config["type"]
+
+            execution_result = {
+                "strategy_id": strategy_id,
+                "strategy_type": strategy_type,
+                "actions_taken": [],
+                "performance_impact": {},
+                "recommendations": [],
+            }
+
+            if strategy_type == "adaptive_test_execution":
+                execution_result.update(
+                    self._execute_adaptive_test_execution(strategy_config, context)
+                )
+            elif strategy_type == "intelligent_test_prioritization":
+                execution_result.update(
+                    self._execute_intelligent_test_prioritization(
+                        strategy_config, context
+                    )
+                )
+            elif strategy_type == "coverage_optimization":
+                execution_result.update(
+                    self._execute_coverage_optimization(strategy_config, context)
+                )
+
+            logger.info("Intelligent testing strategy executed: %s", strategy_id)
+            return execution_result
+
+        except Exception as e:
+            logger.error(f"Failed to execute intelligent testing strategy: {e}")
+            raise
+
+    def _execute_adaptive_test_execution(
+        self, strategy_config: Dict[str, Any], context: Dict[str, Any]
+    ) -> Dict[str, Any]:
+        """Execute adaptive test execution strategy."""
+        return {
+            "actions_taken": ["test_execution_optimization"],
+            "performance_impact": {"execution_time": "optimized"},
+            "recommendations": ["Monitor test execution performance for 15 minutes"],
+        }
+
+    def _execute_intelligent_test_prioritization(
+        self, strategy_config: Dict[str, Any], context: Dict[str, Any]
+    ) -> Dict[str, Any]:
+        """Execute intelligent test prioritization strategy."""
+        return {
+            "actions_taken": ["test_prioritization_optimization"],
+            "performance_impact": {"test_ordering": "optimized"},
+            "recommendations": ["Review test priority assignments"],
+        }
+
+    def _execute_coverage_optimization(
+        self, strategy_config: Dict[str, Any], context: Dict[str, Any]
+    ) -> Dict[str, Any]:
+        """Execute coverage optimization strategy."""
+        return {
+            "actions_taken": ["coverage_optimization"],
+            "performance_impact": {"coverage": "improved"},
+            "recommendations": ["Monitor coverage metrics"],
+        }
