--- tests/test_suite.py
@@ -1,317 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test Suite - Comprehensive Testing for Agent Cellphone V2
-
-This module provides comprehensive testing for all system components
-including launcher modules, utility modules, and configuration management.
-
-Architecture: Single Responsibility Principle - testing orchestration only
-LOC: 150 lines (under 200 limit)
-"""
-
-import sys
-import unittest
-
-from src.utils.stability_improvements import stability_manager, safe_import
-from pathlib import Path
-
-# Add project root to Python path
-project_root = Path(__file__).parent.parent
-sys.path.insert(0, str(project_root))
-
-
-class TestLauncherCore(unittest.TestCase):
-    """Test launcher core functionality"""
-
-    def setUp(self):
-        from src.launchers.launcher_core import LauncherCore
-
-        self.core = LauncherCore()
-
-    def test_initialization(self):
-        """Test launcher core initialization"""
-        self.assertIsNotNone(self.core)
-        self.assertFalse(self.core.system_ready)
-
-    def test_system_check_simulation(self):
-        """Test system check (simulated)"""
-        # This will fail in test environment, but we can test the method exists
-        result = self.core.check_system()
-        self.assertIsInstance(result, bool)
-
-    def test_agent_management(self):
-        """Test agent management methods"""
-        # Test with empty agents
-        status = self.core.get_agent_status()
-        self.assertEqual(status, {})
-
-        # Test shutdown with no agents
-        result = self.core.shutdown_agents()
-        self.assertTrue(result)
-
-
-class TestLauncherModes(unittest.TestCase):
-    """Test launcher modes functionality"""
-
-    def setUp(self):
-        from src.launchers.launcher_modes import LauncherModes
-
-        self.modes = LauncherModes()
-
-    def test_available_modes(self):
-        """Test available modes listing"""
-        modes = self.modes.get_available_modes()
-        expected_modes = ["onboarding", "coordination", "autonomous", "cleanup"]
-
-        for mode in expected_modes:
-            self.assertIn(mode, modes)
-
-    def test_mode_execution_simulation(self):
-        """Test mode execution (simulated)"""
-        test_agents = {}
-
-        # Test all modes with empty agents
-        for mode in self.modes.get_available_modes():
-            result = self.modes.run_mode(mode, test_agents)
-            self.assertIsInstance(result, bool)
-
-
-class TestLauncherCLI(unittest.TestCase):
-    """Test launcher CLI functionality"""
-
-    def setUp(self):
-        from src.launchers.launcher_cli import LauncherCLI
-
-        self.cli = LauncherCLI()
-
-    def test_cli_initialization(self):
-        """Test CLI initialization"""
-        self.assertIsNotNone(self.cli)
-        self.assertIsNotNone(self.cli.parser)
-
-    def test_help_output(self):
-        """Test help output generation"""
-        help_text = self.cli.parser.format_help()
-        self.assertIn("Unified Launcher for Agent Cellphone System", help_text)
-        self.assertIn("--check", help_text)
-        self.assertIn("--init", help_text)
-
-
-class TestFileUtils(unittest.TestCase):
-    """Test file utilities functionality"""
-
-    def setUp(self):
-        from src.utils.file_utils import FileUtils
-
-        self.file_utils = FileUtils()
-        self.test_dir = Path("test_temp")
-        self.test_dir.mkdir(exist_ok=True)
-
-    def tearDown(self):
-        import shutil
-
-        if self.test_dir.exists():
-            shutil.rmtree(self.test_dir)
-
-    def test_directory_creation(self):
-        """Test directory creation"""
-        test_path = self.test_dir / "subdir" / "nested"
-        result = FileUtils.ensure_directory(str(test_path))
-        self.assertTrue(result)
-        self.assertTrue(test_path.exists())
-
-    def test_json_operations(self):
-        """Test JSON read/write operations"""
-        test_file = self.test_dir / "test.json"
-        test_data = {"test": "data", "number": 42}
-
-        # Test write
-        write_result = FileUtils.write_json(str(test_file), test_data)
-        self.assertTrue(write_result)
-        self.assertTrue(test_file.exists())
-
-        # Test read
-        read_data = FileUtils.read_json(str(test_file))
-        self.assertEqual(read_data, test_data)
-
-    def test_file_operations(self):
-        """Test file operations"""
-        test_file = self.test_dir / "test.txt"
-        test_file.write_text("test content")
-
-        # Test file exists
-        self.assertTrue(FileUtils.file_exists(str(test_file)))
-
-        # Test file size
-        size = FileUtils.get_file_size(str(test_file))
-        self.assertIsNotNone(size)
-        self.assertGreater(size, 0)
-
-
-class TestValidationUtils(unittest.TestCase):
-    """Test validation utilities functionality"""
-
-    def setUp(self):
-        from src.utils.validation_utils import ValidationUtils
-
-        self.validation = ValidationUtils()
-
-    def test_email_validation(self):
-        """Test email validation"""
-        valid_emails = ["test@example.com", "user.name@domain.co.uk"]
-        invalid_emails = ["invalid-email", "@domain.com", "user@", "user.domain.com"]
-
-        for email in valid_emails:
-            self.assertTrue(self.validation.is_valid_email(email))
-
-        for email in invalid_emails:
-            self.assertFalse(self.validation.is_valid_email(email))
-
-    def test_url_validation(self):
-        """Test URL validation"""
-        valid_urls = ["https://example.com", "http://test.org/path"]
-        invalid_urls = ["not-a-url", "ftp://example.com", "example.com"]
-
-        for url in valid_urls:
-            self.assertTrue(self.validation.is_valid_url(url))
-
-        for url in invalid_urls:
-            self.assertFalse(self.validation.is_valid_url(url))
-
-    def test_required_fields_validation(self):
-        """Test required fields validation"""
-        data = {"name": "test", "email": "", "age": None}
-        required_fields = ["name", "email", "age", "missing"]
-
-        errors = self.validation.validate_required_fields(data, required_fields)
-
-        self.assertIn("email", errors)
-        self.assertIn("age", errors)
-        self.assertIn("missing", errors)
-        self.assertNotIn("name", errors)
-
-
-class TestSystemUtils(unittest.TestCase):
-    """Test system utilities functionality"""
-
-    def setUp(self):
-        from src.utils.system_utils import SystemUtils
-
-        self.system_utils = SystemUtils()
-
-    def test_system_info(self):
-        """Test system information gathering"""
-        info = self.system_utils.get_system_info()
-
-        self.assertIn("platform", info)
-        self.assertIn("python_version", info)
-        self.assertIn("current_working_directory", info)
-
-    def test_dependencies_check(self):
-        """Test dependencies checking"""
-        deps = self.system_utils.check_dependencies()
-
-        # Built-in modules should always be available
-        self.assertTrue(deps["pathlib"])
-        self.assertTrue(deps["logging"])
-        self.assertTrue(deps["json"])
-        self.assertTrue(deps["re"])
-
-    def test_logging_setup(self):
-        """Test logging setup"""
-        result = self.system_utils.setup_logging("INFO")
-        self.assertTrue(result)
-
-
-class TestConfigManager(unittest.TestCase):
-    """Test configuration manager functionality"""
-
-    def setUp(self):
-        from src.utils.config_manager import ConfigManager
-
-        self.config_manager = ConfigManager("test_config.yaml")
-
-    def test_config_validation(self):
-        """Test configuration validation"""
-        # Test with empty config
-        errors = self.config_manager.validate_config()
-        self.assertIn("system", errors)
-        self.assertIn("agents", errors)
-
-    def test_config_operations(self):
-        """Test configuration operations"""
-        # Test setting and getting config values
-        success = self.config_manager.set_config("test.key", "test_value")
-        self.assertTrue(success)
-
-        value = self.config_manager.get_config("test.key")
-        self.assertEqual(value, "test_value")
-
-
-def run_all_tests():
-    """Run all test suites"""
-    print("🧪 Running comprehensive test suite...")
-
-    # Create test suite
-    test_suite = unittest.TestSuite()
-
-    # Add test classes
-    test_classes = [
-        TestLauncherCore,
-        TestLauncherModes,
-        TestLauncherCLI,
-        TestFileUtils,
-        TestValidationUtils,
-        TestSystemUtils,
-        TestConfigManager,
-    ]
-
-    for test_class in test_classes:
-        tests = unittest.TestLoader().loadTestsFromTestCase(test_class)
-        test_suite.addTests(tests)
-
-    # Run tests
-    runner = unittest.TextTestRunner(verbosity=2)
-    result = runner.run(test_suite)
-
-    # Print summary
-    print(f"\n📊 Test Results Summary:")
-    print(f"  Tests run: {result.testsRun}")
-    print(f"  Failures: {len(result.failures)}")
-    print(f"  Errors: {len(result.errors)}")
-
-    if result.failures:
-        print(f"\n❌ Failures:")
-        for test, traceback in result.failures:
-            print(f"  {test}: {traceback}")
-
-    if result.errors:
-        print(f"\n❌ Errors:")
-        for test, traceback in result.errors:
-            print(f"  {test}: {traceback}")
-
-    success = len(result.failures) == 0 and len(result.errors) == 0
-    print(f"\n{'✅ All tests passed!' if success else '❌ Some tests failed'}")
-
-    return success
-
-
-def main():
-    """Main entry point for test suite"""
-    import argparse
-
-    parser = argparse.ArgumentParser(description="Test Suite CLI")
-    parser.add_argument("--run", action="store_true", help="Run all tests")
-    parser.add_argument("--verbose", "-v", action="store_true", help="Verbose output")
-
-    args = parser.parse_args()
-
-    if args.run:
-        success = run_all_tests()
-        sys.exit(0 if success else 1)
-    else:
-        parser.print_help()
-
-
-if __name__ == "__main__":
-    main()
